Further work done

+++ src/services/run.js
@@ -7,20 +7,92 @@ const errorUtil = require(path.resolve(global.utilsFolder, "error"));
 const dateUtil = require(path.resolve(global.utilsFolder, "date"));
 
 const acceptedActionsTypes = {
-    START: "START",
-    MOVE_KART: "MOVE_KART",
-    CONTINUE_WORKING: "CONTINUE_WORKING",
-    KILL: "KILL",
-    END_RUN: "END_RUN"//TODO NELSON
+    //TODO NELSON the run needs do have the status as "RUNNING" before this
+    //TODO NELSON SHOULD not have an iteration at this moment
+    //TODO NELSON socket events should be blocked for this project before this(I think we can see this if it does not exist an iteration at the moment(with a start time only))
+    //TODO NELSON create a new iteration(with a start time)
+    //TODO NELSON run total time starts decreasing(add start time to the run)
+    //TODO NELSON set status of the run to running, set status of the project to running
+    //TODO NELSON unlock socket events for the project
+    START: {
+        KEY: "START",
+        canExecute: function (project, run, iteration, measurements) {
+            return (project.status === "CREATED" &&
+                run.status === "RUNNING" &&
+                iteration === null &&
+                Array.isArray(measurements) &&
+                measurements.length === 0);
+        }
+    },
+    //TODO NELSON can only happen when all stations have completed(current iteration has start but no stop time)
+    //TODO NELSON also all measurements for that iteration must have a start and stop time(meaning all stations have stopped work for that iteration)
+    //TODO then mark iteration as done(add a stop time)
+    //TODO NELSON block socket events for project(meaning it does not exist an iteration with a start time only)
+    MOVE_KART: {
+        KEY: "MOVE_KART",
+        canExecute: function (project, run, iteration, measurements) {
+            let baseCheck = (project.status === "RUNNING" &&
+                run.status === "RUNNING" &&
+                iteration &&
+                iteration.startTime &&
+                !iteration.stopTime &&
+                Array.isArray(measurements) &&
+                measurements.length === project.numStations);
+            if(baseCheck) {
+                let result = true;
+                measurements.forEach(function(measurement) {
+                    //each measurement must have a startTime and stopTime to be able to move kart
+                    if(!measurement.startTime || !measurement.stopTime) {
+                        result = false;
+                        return;
+                    }
+                });
+                return result;
+            } else {
+                return false;
+            }
+        }
+    },
+    //TODO NELSON this action can only be executed after move kart is clicked(meaning when it does not exist an iteration with a start time only related to that run)
+    //TODO NELSON unlocks socket events for project
+    //TODO NELSON create a new iteration for the run in the DB(add only a start time to it)
+    CONTINUE_WORKING: {
+        KEY: "CONTINUE_WORKING",
+        canExecute: function (project, run, iteration, measurements) {
+            return (project.status === "RUNNING" &&
+                run.status === "RUNNING" &&
+                iteration &&
+                !iteration.startTime &&
+                !iteration.stopTime &&
+                Array.isArray(measurements) &&
+                measurements.length === 0);
+        }
+    },
+    //TODO NELSON -> destroy the project and info associated to it(runs, iterations, measurements and events)
+    //TODO NELSON -> maybe mark the project with a status of messy and redirect to the project page(then the user can delete the project there)
+    KILL: {
+        KEY: "KILL",
+        canExecute: function (project, run, iteration, measurements) {
+            return true;//TODO NELSON think about this, but I think a project can be killed without any check at the moment
+        }
+    },
+    //TODO NELSON -> Frontend needs to call this action when the run total time reached zero
+    END_RUN: {
+        KEY: "END_RUN",
+        canExecute: function (project, run, iteration, measurements) {
+            return (project.status === "RUNNING" &&
+                run.status === "RUNNING" &&
+                run.startTimestamp &&
+                run.startTimestamp + (run.totalTime * 60) <= dateUtil.getCurrentTimestamp());
+        }
+    }
 };
 
 
-const parseAction = async function (project, run, iteration, measurements, actionType) {
-    //TODO NELSON I might need to receive iteration as param here (latest iteration related to the run(if it is null it can only fit in the start action)??? -> i think it works for all cases here)
-    //TODO NELSON I might also need the measurements associated to the current iteration
+const executeRunAction = async function (project, run, iteration, measurements, actionType) {
     switch (actionType) {
-        case acceptedActionsTypes.START:
-            if (project.status === "CREATED" && run.status === "RUNNING" && iteration === null && Array.isArray(measurements) && measurements.length === 0) {
+        case acceptedActionsTypes.START.KEY:
+            if (acceptedActionsTypes.START.canExecute(project, run, iteration, measurements)) {
                 //TODO NELSON implement transactions here
                 let runStartTime = dateUtil.getCurrentTimestamp();
                 let promises = [
@@ -33,7 +105,7 @@ const parseAction = async function (project, run, iteration, measurements, actio
                     return run.createNewIterationForRun(previousIterationNumber, createWithAStartTime).then(function (createdIteration) {
                         project.status = "RUNNING";
                         run.startTimestamp = runStartTime;
-                        return  {
+                        return {
                             project: project,
                             run: run,
                             iteration: createdIteration,
@@ -47,17 +119,9 @@ const parseAction = async function (project, run, iteration, measurements, actio
             } else {
                 errorUtil.createAndThrowGenericError("Cannot start a new run", 400);
             }
-            //TODO NELSON the run needs do have the status as "RUNNING" before this
-            //TODO NELSON SHOULD not have an iteration at this moment
-            //TODO NELSON socket events should be blocked for this project before this(I think we can see this if it does not exist an iteration at the moment(with a start time only))
-            //TODO NELSON create a new iteration(with a start time)
-            //TODO NELSON run total time starts decreasing(add start time to the run)
-            //TODO NELSON set status of the run to running, set status of the project to running
-            //TODO NELSON unlock socket events for the project
             break;
-            //return null;
-        case acceptedActionsTypes.MOVE_KART:
-            if (project.status === "RUNNING" && run.status === "RUNNING" && iteration && iteration.startTime && !iteration.stopTime && Array.isArray(measurements) && measurements.length === project.numStations /* && TODO NELSON cada measure tem de ter um start e stop time */) {
+        case acceptedActionsTypes.MOVE_KART.KEY:
+            if (acceptedActionsTypes.MOVE_KART.canExecute(project, run, iteration, measurements)) {
                 //TODO NELSON implement transactions here
                 let iterationStopTime = dateUtil.getCurrentTimestamp();
                 let promises = [
@@ -67,7 +131,7 @@ const parseAction = async function (project, run, iteration, measurements, actio
                     let previousIterationNumber = iteration.number;
                     let createWithAStartTime = false;
                     return run.createNewIterationForRun(previousIterationNumber, createWithAStartTime).then(function (createdIteration) {
-                        return  {
+                        return {
                             project: project,
                             run: run,
                             iteration: createdIteration,
@@ -81,14 +145,9 @@ const parseAction = async function (project, run, iteration, measurements, actio
             } else {
                 errorUtil.createAndThrowGenericError("Cannot move kart", 400);
             }
-            //TODO NELSON can only happen when all stations have completed(current iteration has start but no stop time)
-            //TODO NELSON also all measurements for that iteration must have a start and stop time(meaning all stations have stopped work for that iteration)
-            //TODO then mark iteration as done(add a stop time)
-            //TODO NELSON block socket events for project(meaning it does not exist an iteration with a start time only)
-            //break;
-            return null;
-        case acceptedActionsTypes.CONTINUE_WORKING:
-            if (project.status === "RUNNING" && run.status === "RUNNING" && iteration && !iteration.startTime && !iteration.stopTime && Array.isArray(measurements) && measurements.length === 0) {
+            break;
+        case acceptedActionsTypes.CONTINUE_WORKING.KEY:
+            if (acceptedActionsTypes.CONTINUE_WORKING.canExecute(project, run, iteration, measurements)) {
                 let iterationStartTime = dateUtil.getCurrentTimestamp();
                 return iteration.updateOne({startTime: iterationStartTime}).then(function (updatedIteration) {
                     return {
@@ -102,31 +161,28 @@ const parseAction = async function (project, run, iteration, measurements, actio
             } else {
                 errorUtil.createAndThrowGenericError("Cannot continue working", 400);
             }
-
-            //TODO NELSON this action can only be executed after move kart is clicked(meaning when it does not exist an iteration with a start time only related to that run)
-            //TODO NELSON unlocks socket events for project
-            //TODO NELSON create a new iteration for the run in the DB(add only a start time to it)
-            //break;
-            return null;
-        case acceptedActionsTypes.KILL:
-            //TODO NELSON -> destroy the project and info associated to it(runs, iterations, measurements and events)
-            //TODO NELSON -> maybe mark the project with a status of messy and redirect to the project page(then the user can delete the project there)
-            return projectModel.deleteProjectByNumber(project.number).then(function (deletedProject) {
-                return {
-                    project: deletedProject,
-                    run: run,
-                    iteration: iteration,
-                    measurements: measurements,
-                    actionType: actionType
-                };
-            });
-            //break;
-        case acceptedActionsTypes.END_RUN:
-            if (project.status === "RUNNING" && run.status === "RUNNING" && run.startTimestamp && run.startTimestamp + (run.totalTime * 60) <= dateUtil.getCurrentTimestamp()) {
+            break;
+        case acceptedActionsTypes.KILL.KEY:
+            if (acceptedActionsTypes.KILL.canExecute(project, run, iteration, measurements)) {
+                return projectModel.deleteProjectByNumber(project.number).then(function (deletedProject) {
+                    return {
+                        project: deletedProject,
+                        run: run,
+                        iteration: iteration,
+                        measurements: measurements,
+                        actionType: actionType
+                    };
+                });
+            } else {
+                errorUtil.createAndThrowGenericError("Cannot Kill project", 400);
+            }
+            break;
+        case acceptedActionsTypes.END_RUN.KEY:
+            if (acceptedActionsTypes.END_RUN.canExecute(project, run, iteration, measurements)) {
                 let promises = [
                     run.updateOne({status: "FINISHED"})
                 ];
-                if(run.number === project.numRuns) {
+                if (run.number === project.numRuns) {
                     project.status = "FINISHED";
                     promises.push(project.updateOne({status: "FINISHED"}));
                 }
@@ -145,9 +201,7 @@ const parseAction = async function (project, run, iteration, measurements, actio
             } else {
                 errorUtil.createAndThrowGenericError("Cannot End run yet", 400);
             }
-            //TODO NELSON -> Frontend needs to call this action when the run total time reached zero
-            //break;
-            return null;
+            break;
         default:
             return null;
         //TODO NELSON think how to handle the default case(throw an error or simply return as null)
@@ -170,7 +224,7 @@ module.exports.update = async function (project, run, iteration, measurements, a
         let {actionType} = actionData;
         if (acceptedActionsTypes.hasOwnProperty(actionType)) {
             //TODO NELSON validate here the accepted action types
-            return parseAction(project, run, iteration, measurements, actionType);
+            return executeRunAction(project, run, iteration, measurements, actionType);
         } else {
             errorUtil.createAndThrowGenericError("Invalid action type", 400);
         }
